feat: change app settings responsively
