+++ apps/mobile/app/_layout.tsx
@@ -1,48 +1,18 @@
 import { AuthProvider } from '@/context/AuthContext';
-import { SSEProvider } from '@/context/SSEContext';
 import '@/locales/i18n';
-import { Slot } from 'expo-router';
-import { StatusBar } from 'expo-status-bar';
-import { DarkTheme as NavDarkTheme, DefaultTheme as NavLightTheme, ThemeProvider } from '@react-navigation/native';
-import { adaptNavigationTheme, PaperProvider } from 'react-native-paper';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
+import { AppSettingProvider } from '@/context/AppSettingContext';
+import { AppTheme } from '@/components/AppTheme';
 import '../global.css';
-import { Surface } from '@/components/common';
-import Themes from '@/constants/Themes';
-import { useAppSetting } from '@/hooks/useAppSetting';
 
 export default function Root() {
-  const { isSettingLoading, effectiveColorScheme, currentColor } = useAppSetting();
-
-  if (isSettingLoading) {
-    return null;
-  }
-
-  const paperTheme = Themes[effectiveColorScheme ?? 'light'][currentColor];
-
-  const { DarkTheme, LightTheme } = adaptNavigationTheme({
-    reactNavigationDark: NavDarkTheme,
-    reactNavigationLight: NavLightTheme,
-    materialDark: Themes.dark.default,
-    materialLight: Themes.light.default
-  });
-
-  const statusBarStyle = effectiveColorScheme === 'dark' ? 'light' : 'dark';
-
   return (
     <SafeAreaProvider>
-      <AuthProvider>
-        <PaperProvider theme={paperTheme}>
-          <ThemeProvider value={effectiveColorScheme === 'dark' ? DarkTheme : LightTheme}>
-            <SSEProvider>
-              <Surface style={{ flex: 1 }} testID="root-surface">
-                <StatusBar style={statusBarStyle} />
-                <Slot />
-              </Surface>
-            </SSEProvider>
-          </ThemeProvider>
-        </PaperProvider>
-      </AuthProvider>
+      <AppSettingProvider>
+        <AuthProvider>
+          <AppTheme />
+        </AuthProvider>
+      </AppSettingProvider>
     </SafeAreaProvider>
   );
 }
